feat(BlazorPortfolio.Client): Create PortfolioService
Created a new service class called PortfolioService in the BlazorPortfolio.Client project. This service will be responsible for fetching and managing portfolio data.
The PortfolioService will handle tasks such as reading the portfolio data from the JSON file, providing methods to access and manipulate the data, and potentially interacting with external APIs or data sources in the future.